+++ src/main/scala/com/raquo/airstream/features/InternalTryObserver.scala
@@ -4,7 +4,7 @@ import com.raquo.airstream.core.{InternalObserver, Transaction}
 
 import scala.util.{Failure, Success}
 
-trait InternalTryObserver[A] extends InternalObserver[A] {
+trait InternalTryObserver[-A] extends InternalObserver[A] {
 
   override protected[airstream] final def onNext(nextValue: A, transaction: Transaction): Unit = {
     onTry(Success(nextValue), transaction)