+++ src/test/scala/com/raquo/airstream/errors/ObserverErrorSpec.scala
@@ -23,14 +23,19 @@ class ObserverErrorSpec extends FunSpec with Matchers with BeforeAndAfter {
   val err3 = new Exception("err3")
   val err31 = new Exception("err31")
 
-  AirstreamError.registerUnhandledErrorCallback(errorEffects += Effect("unhandled", _))
+  private val errorCallback = (err: Throwable) => {
+    errorEffects += Effect("unhandled", err)
+    ()
+  }
 
   before {
+    AirstreamError.registerUnhandledErrorCallback(errorCallback)
     AirstreamError.unregisterUnhandledErrorCallback(AirstreamError.consoleErrorCallback)
   }
 
   after {
     AirstreamError.registerUnhandledErrorCallback(AirstreamError.consoleErrorCallback)
+    AirstreamError.unregisterUnhandledErrorCallback(errorCallback)
     calculations.clear()
     effects.clear()
     errorEffects.clear()